Conflicts:
	db/schema.rb

+++ app/helpers/arguments_helper.rb
@@ -41,7 +41,7 @@ def argument_items(argument)
     link_items << link_item(t('destroy'), argument_path(argument, destroy: true), data: {confirm: t('destroy_confirmation'), method: 'delete', 'skip-pjax' => 'true'}, fa: 'close')
   end
   if active_for_user?(:notifications, current_user)
-    if current_profile.following?(argument)
+    if current_profile && current_profile.following?(argument)
       link_items << link_item(t('forums.unfollow'), follows_path(argument_id: argument.id), fa: 'bell-slash', data: {method: 'delete', 'skip-pjax' => 'true'})
     else
       link_items << link_item(t('forums.follow'), follows_path(argument_id: argument.id), fa: 'bell', data: {method: 'create', 'skip-pjax' => 'true'})
